Refactor client_creds tests and add PPE coverage
